restructure engine module
