+++ nipype/caching/memory.py
@@ -21,7 +21,7 @@ import glob
 
 from ..interfaces.base import BaseInterface
 from ..pipeline.engine import Node
-from ..pipeline.utils import modify_paths
+from ..pipeline.engine.utils import modify_paths
 
 ################################################################################
 # PipeFunc object: callable interface to nipype.interface objects
