+++ nipype/pipeline/plugins/base.py
@@ -21,10 +21,11 @@ import numpy as np
 import scipy.sparse as ssp
 
 
-from ..utils import (nx, dfs_preorder, topological_sort)
-from ..engine import (MapNode, str2bool)
+from ...utils.filemanip import savepkl, loadpkl
+from ...utils.misc import str2bool
+from ..engine.utils import (nx, dfs_preorder, topological_sort)
+from ..engine import MapNode
 
-from nipype.utils.filemanip import savepkl, loadpkl
 
 from ... import logging
 logger = logging.getLogger('workflow')
